Update to v2.20. Resolve purchases API fields

+++ shipping_addresses.go
@@ -35,7 +35,7 @@ type ShippingAddressesService interface {
 // ShippingAddress represents a shipping address
 type ShippingAddress struct {
 	XMLName   xml.Name `xml:"shipping_address"`
-	ID        int      `xml:"id,omitempty"`
+	ID        int64    `xml:"id,omitempty"`
 	FirstName string   `xml:"first_name"`
 	LastName  string   `xml:"last_name"`
 	Nickname  string   `xml:"nickname,omitempty"`
